requirements.txt
# No additional dependencies are needed for this simple function
#google-generativeai==0.3.2
google-genai